requirements.txt
numpy==1.19.5
PyYAML==6.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
sklearn==0.0
stanza==1.1.1
tqdm==4.62.3
pyrouge==0.1.3